Completed audio augmentation

+++ src/data_augmentation/utils.py
@@ -15,9 +15,35 @@ class WhenAlreadyDone(Enum):
     OVERWRITE = 1
     SKIP = 2
 
+#---------- Enum for how much to equalize num samples per species -----------
+
+# How many augmentations to create 
+# for each species. Measured against
+# the number of samples of the species 
+# with the most number of available
+# samples.
+# Meaning: 
+#    TENTH: all species will have at least a 				  
+#           10th of samples in the most populous 
+#           species 	  
+#   MEDIAN: all species will have at least the median
+#    	    number of samples in the species populations
+#      MAX: all species will end up with the number of
+#           species of the most populously represented species
+#           in the training set: 
+class AugmentationGoals(Enum):
+    TENTH   = 0
+    MEDIAN  = 1
+    MAX     = 2
+    NUMBER  = 3 # not implemented yet: Arbitrary number of samples to create
+
 #------------------------------ Utility  -------------
 
 class Utils:
+
+    #------------------------------------
+    # noise_multiplier
+    #-------------------
 
     @classmethod
     def noise_multiplier(cls, orig_recording, noise):
@@ -27,6 +53,10 @@ class Utils:
         orig_rms  = np.sqrt(np.mean(orig_recording**2))
         desired_rms = orig_rms / (10 ** (float(snr) / 20))
         return desired_rms / noise_rms
+    
+    #------------------------------------
+    # create_folder 
+    #-------------------
     
     @classmethod
     def create_folder(cls, dir_path, overwrite_policy):
@@ -52,6 +82,8 @@ class Utils:
         :rtype: bool
         
         '''
+        if not isinstance(overwrite_policy, WhenAlreadyDone):
+            raise TypeError(f"Overwrite policy must be a member of WhenAlreadyDone, not {type(overwrite_policy)}") 
         if os.path.exists(dir_path):
             if overwrite_policy == WhenAlreadyDone.OVERWRITE:
                 shutil.rmtree(dir_path)
@@ -70,7 +102,11 @@ class Utils:
         else:
             os.mkdir(dir_path)
             return True
-        
+
+    #------------------------------------
+    # sample_compositions_by_species 
+    #-------------------
+
     @classmethod
     def sample_compositions_by_species(cls, path, augmented):
         num_samples_in = {} # initialize dict - usage num_samples_in['CORALT_S'] = 64
@@ -85,8 +121,54 @@ class Utils:
                 num_samples_in[species]= aug_type_dict
             else:
                 num_samples_in[species] = {"num_samples":len(os.listdir(os.path.join(path, species)))}
-        return pd.DataFrame.from_dict(num_samples_in, orient='index').sort_index()
+        df = pd.DataFrame.from_dict(num_samples_in, orient='index').sort_index()
+        # Add a nice col header:
+        df.colums = ['num_samples']
+        df.colums
+        return df
     
+    #------------------------------------
+    # find_species_names 
+    #-------------------
+    
+    @classmethod
+    def find_species_names(cls, species_root):
+        '''
+        Given the root of a directory in which only
+        subdirectories exist that are named after a 
+        bird species, return a list of those species
+        names.
+        
+        Expectation for species_root:
+        
+                    species_root
+                    
+            brd1_dir   some_file   brd2_dir, brd3_dir ...
+              A   N   Y   T    H    I    N    G
+              
+        I.e. there may be files in species_root. But anything
+        that is a directory must have a species name.
+        
+        :param species_root: root of bird species related subdirectories
+        :type species_root: str
+        :return: list of species names
+        :rtype: [str]
+        '''
+        try:
+            species_names = [dir_name 
+                             for dir_name 
+                             in os.listdir(species_root)
+                             if os.path.isdir(os.path.join(species_root, dir_name))
+                             ]
+        except FileNotFoundError:
+            return []
+        return species_names
+
+    
+    #------------------------------------
+    # find_total_recording_length
+    #-------------------
+
     @classmethod
     def find_total_recording_length(cls, species_dir_path):
         total_duration = 0
@@ -94,6 +176,11 @@ class Utils:
             y, sr = librosa.load(os.path.join(species_dir_path, recording))
             total_duration += librosa.get_duration(y, sr)
         return total_duration
+
+    #------------------------------------
+    # recording_lengths_by_species
+    #-------------------
+
     
     @classmethod
     def recording_lengths_by_species(cls, path):
@@ -103,6 +190,10 @@ class Utils:
             num_samples_in[species] = {"total_recording_length": rec_len} 
         return pd.DataFrame.from_dict(num_samples_in, orient='index').sort_index()
 
+    #------------------------------------
+    # count_max_augs 
+    #-------------------
+
     @classmethod
     def count_max_augs(cls, distribution):
         count = 0
@@ -110,7 +201,11 @@ class Utils:
             if val != 0:
                 count += 1
         return count
-    
+
+    #------------------------------------
+    # find_in_dir_tree 
+    #-------------------
+
     @classmethod
     def find_in_dir_tree(cls, root_dir, pattern='*', entry_type='file'):
         '''
@@ -175,3 +270,93 @@ class Utils:
                     file_list.append(os.path.join(dName, fname))
     
         return file_list
+
+    #------------------------------------
+    # compute_num_augs_per_species
+    #-------------------
+
+    @classmethod
+    def compute_num_augs_per_species(cls, aug_volumes, sample_distrib_df):
+        '''
+        Return a dict mapping species name to 
+        number of samples that should be available after
+        augmentation. 
+        
+        The aug_volumes arg is either a dict mapping species name
+        to an AugmentationGoals (TENTH, MEDIAN, MAX), or just
+        an individual AugmentationGoals. The augmentation volume 
+        indicates to which degree the least populated species should 
+        be filled with augmented samples.  
+
+        The sample_distrib_df is a dataframe whose row labels are 
+        species names and the single column's values are numbers
+        of available samples for training/validation/test for the
+        respective row's species:
+        
+                        'num_samples'
+               'bird1'   2
+               'bird2'   6
+        
+        :param aug_volumes: how many augmentations for each species
+        :type aug_volumes: {AugmentationGoals | {str : AugmentationGoals}}
+        :param sample_distrib_df: distribution of initially available
+            sample numbers for each species
+        :type sample_distrib_df: pandas.DataFrame
+        :return: dict mapping each species to the 
+            number of samples that need to be created.
+        :rtype: {str : int}
+        '''
+        
+        # For convenience:
+        df = sample_distrib_df
+        species_names = df.index
+        
+        # Find row index of species with maximum number of recordings:
+        max_recs_idx = df['num_samples'].argmax()
+        
+        # Pull that row out of the df, getting a series:
+        #  num_samples    6
+        #  Name: HENLES_S, dtype: int6
+        row_of_max = df.iloc[[max_recs_idx]].iloc[0]
+        _max_species = row_of_max.name
+        max_num_samples = row_of_max.num_samples
+        
+        # Compute the various methods (tenth, median, etc), 
+        # relative to the max num of recordings: 
+        # Plus 1 to avoid asking for 0 augmentations:
+        tenth_max_num_samples = 1 + max_num_samples//10
+        
+        # Note: corner case for following statement:
+        #   one species with no samples, another with 
+        #   1 sample. Median is 0. That's fine, b/c
+        #   we cannot augment the 0-samples anyway:
+        median_num_samples = int(df['num_samples'].median())
+        
+        # Mapping goal names to number of recordings
+        # that should be available after audio augmentation:
+        volumes = {AugmentationGoals.TENTH  : tenth_max_num_samples,
+                   AugmentationGoals.MEDIAN : median_num_samples,
+                   AugmentationGoals.MAX    : max_num_samples
+                   }
+        
+        aug_requirements = {}
+        if isinstance(aug_volumes, AugmentationGoals):
+            # Caller passed one aug volume for all species,
+            # rather than a dict w/ species-by-species.
+            # Create a dict with all species having that
+            # one goal. That way all cases can be treated 
+            # the same below:
+            aug_volumes = {species_name : aug_volumes
+                           for species_name
+                           in species_names 
+                           }
+        # Do the computation:
+        # Have dict of species-name : AugmentationGoals:
+        for species in species_names:
+            # ******* TODO: Ability to specify number manually:
+            aug_requirement = aug_volumes[species]
+            end_goal = volumes[aug_requirement]
+            curr_num_recordings = df.loc[species, 'num_samples']
+            aug_requirements[species] = max(0,end_goal - curr_num_recordings)
+        
+        return aug_requirements
